fix: move plugins def into plugins folder
