--- definitions/plugins/astroboy-view/app.ts
@@ -1,3 +1,0 @@
-export interface IAstroboyViewPluginAppExtends {
-  readonly view: any;
-}
