+++ plugins/astroboy-router/contract.ts
@@ -1,4 +1,4 @@
-import { IMiddlewareOptions } from '../../config';
+import { IMiddlewareOptions } from '../../definitions/config';
 
 export interface IAstroboyRouterPluginMiddlewareExtends {
   'astroboy-router': IMiddlewareOptions<any>;
