+++ plugins/astroboy-static/contract.ts
@@ -1,4 +1,4 @@
-import { IMiddlewareOptions } from '../../config';
+import { IMiddlewareOptions } from '../../definitions/config';
 
 export interface IAstroboyStaticPluginMiddlewareExtends {
   'astroboy-static': IMiddlewareOptions<{ root?: string }>;
